Added is_download_complete flag in metadata file

+++ src/destination/destination_adapter.cpp
@@ -20,17 +20,25 @@ bool DestinationAdapter::UploadFiles(std::string download_folder, JsonHandler& j
 
         FileMetadata file_metadata = FileMetadataUtils::GetFileMetadataOfLocalFile(itr.path().filename().string());
 
+        if(!FileMetadataUtils::IsCompletelyDownloaded(file_metadata.unique_id))
+            continue;
+
         bool success = destination_type->UploadFile(file_metadata);
 
         if(success)
         {
+            std::cerr << "Info: File uploaded successfully: " << std::endl;
+            std::cerr << "Info: Local file name: " << file_metadata.local_path << std::endl;
+            std::cerr << "Info: Actual file name: " << file_metadata.name << std::endl;
             FileMetadataUtils::UpdateFileMetadataForUpload(file_metadata.unique_id, file_metadata.name);
             json_handler.SaveToFile(FileMetadataUtils::file_id_to_metadata);
             fs::remove(itr.path());
         }
         else
         {
-            std::cerr << "Error: Failed to upload file: " << itr.path().filename().string() << std::endl;
+            std::cerr << "Error: Failed to upload file: " << std::endl;
+            std::cerr << "Info: Local file name: " << file_metadata.local_path << std::endl;
+            std::cerr << "Info: Actual file name: " << file_metadata.name << std::endl;
             all_success = false;
         }
     }
